requirements.txt
absl-py==0.7.1
astor==0.8.0
bottle==0.12.17
certifi==2019.6.16
chardet==3.0.4
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.22.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
pkg-resources==0.0.0
protobuf==3.9.0
requests==2.22.0
rocket-python==1.2.8
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
mattermostdriver==7.0.1
